ChatCrud
ChatVeiw and ChatForm have been combined into ChatCrud. This will make it easier to get the UserView's message button to work
