--- packages/ui/src/views/objects/chat/ChatUpsert/ChatUpsert.tsx
@@ -1,94 +1,0 @@
-import { Chat, ChatCreateInput, ChatUpdateInput, endpointGetChat, endpointPostChat, endpointPutChat } from "@local/shared";
-import { fetchLazyWrapper } from "api";
-import { MaybeLargeDialog } from "components/dialogs/LargeDialog/LargeDialog";
-import { TopBar } from "components/navigation/TopBar/TopBar";
-import { SessionContext } from "contexts/SessionContext";
-import { Formik } from "formik";
-import { ChatForm, chatInitialValues, transformChatValues, validateChatValues } from "forms/ChatForm/ChatForm";
-import { useFormDialog } from "hooks/useFormDialog";
-import { useObjectFromUrl } from "hooks/useObjectFromUrl";
-import { useUpsertActions } from "hooks/useUpsertActions";
-import { useContext } from "react";
-import { useTranslation } from "react-i18next";
-import { toDisplay } from "utils/display/pageTools";
-import { PubSub } from "utils/pubsub";
-import { ChatShape } from "utils/shape/models/chat";
-import { ChatUpsertProps } from "../types";
-
-export const ChatUpsert = ({
-    isCreate,
-    isOpen,
-    onCancel,
-    onCompleted,
-    overrideObject,
-}: ChatUpsertProps) => {
-    const { t } = useTranslation();
-    const session = useContext(SessionContext);
-    const display = toDisplay(isOpen);
-
-    const { isLoading: isReadLoading, object: existing } = useObjectFromUrl<Chat, ChatShape>({
-        ...endpointGetChat,
-        objectType: "Chat",
-        overrideObject,
-        transform: (data) => chatInitialValues(session, data),
-    });
-
-    const {
-        fetch,
-        handleCancel,
-        handleCompleted,
-        isCreateLoading,
-        isUpdateLoading,
-    } = useUpsertActions<Chat, ChatCreateInput, ChatUpdateInput>({
-        display,
-        endpointCreate: endpointPostChat,
-        endpointUpdate: endpointPutChat,
-        isCreate,
-        onCancel,
-        onCompleted,
-    });
-    const { formRef, handleClose } = useFormDialog({ handleCancel });
-
-    return (
-        <MaybeLargeDialog
-            display={display}
-            id="chat-upsert-dialog"
-            isOpen={isOpen ?? false}
-            onClose={handleClose}
-        >
-            <TopBar
-                display={display}
-                onClose={handleClose}
-                title={t(isCreate ? "CreateChat" : "UpdateChat")}
-            />
-            <Formik
-                enableReinitialize={true}
-                initialValues={existing}
-                onSubmit={(values, helpers) => {
-                    if (!isCreate && !existing) {
-                        PubSub.get().publishSnack({ messageKey: "CouldNotReadObject", severity: "Error" });
-                        return;
-                    }
-                    fetchLazyWrapper<ChatCreateInput | ChatUpdateInput, Chat>({
-                        fetch,
-                        inputs: transformChatValues(values, existing, isCreate),
-                        onSuccess: (data) => { handleCompleted(data); },
-                        onCompleted: () => { helpers.setSubmitting(false); },
-                    });
-                }}
-                validate={async (values) => await validateChatValues(values, existing, isCreate)}
-            >
-                {(formik) => <ChatForm
-                    display={display}
-                    isCreate={isCreate}
-                    isLoading={isCreateLoading || isReadLoading || isUpdateLoading}
-                    isOpen={true}
-                    onCancel={handleCancel}
-                    onClose={handleClose}
-                    ref={formRef}
-                    {...formik}
-                />}
-            </Formik>
-        </MaybeLargeDialog>
-    );
-};
