+++ packages/ui/src/views/objects/chat/index.ts
@@ -1,1 +1,1 @@
-export * from "./ChatUpsert/ChatUpsert";
+export * from "./ChatCrud/ChatCrud";
